[PS2SQEXT] Update Psf2SqExtractorWorker to use non-.exe method for PSF2 unpacking.

+++ VGMToolbox/tools/xsf/Psf2SqExtractorWorker.cs
@@ -43,9 +43,6 @@ namespace VGMToolbox.tools.xsf
             string fileName;
             string outputDir;
             string libOutputDir;
-            
-            string unpkOutput;
-            string unpkError;
 
             string formatString = XsfUtil.GetXsfFormatString(pPath);
 
@@ -55,59 +52,31 @@ namespace VGMToolbox.tools.xsf
                 fileDir = Path.GetDirectoryName(filePath);
                 fileName = Path.GetFileNameWithoutExtension(filePath);
                 
-                outputDir = XsfUtil.UnpackPsf2(filePath, out unpkOutput, out unpkError);
+                outputDir = XsfUtil.UnpackPsf2(filePath);
                 
                 // parse ini
                 iniFiles = Directory.GetFiles(outputDir, "PSF2.INI", SearchOption.AllDirectories);
-                using (FileStream iniFs = File.Open(iniFiles[0], FileMode.Open, FileAccess.Read))
-                {
-                    // parse ini file to get SQ info
-                    psf2IniStruct = Psf2.ParseClsIniFile(iniFs);
-                }
 
-                using (FileStream fs = File.OpenRead(pPath))
+                if (iniFiles.Length > 0)
                 {
-                    Psf2 psf2File = new Psf2();
-                    psf2File.Initialize(fs, pPath);
-                    
-                    // check for libs
-                    libPaths = psf2File.GetLibPathArray();
-                }
-                if ((libPaths == null) || (libPaths.Length == 0)) // PSF2
-                {
-                    // copy the SQ file out (should only be one)
-                    sqFiles = Directory.GetFiles(outputDir, psf2IniStruct.SqFileName, SearchOption.AllDirectories);
-
-                    if (sqFiles.Length > 0)
+                    using (FileStream iniFs = File.Open(iniFiles[0], FileMode.Open, FileAccess.Read))
                     {
-                        if (!String.IsNullOrEmpty(psf2IniStruct.SequenceNumber))
-                        {
-                            outputSqFileName = String.Format("{0}_n={1}.SQ", outputDir, psf2IniStruct.SequenceNumber);
-                        }
-                        else
-                        {
-                            outputSqFileName = String.Format("{0}.SQ", outputDir);                                        
-                        }
-                        File.Copy(sqFiles[0], outputSqFileName, true);
+                        // parse ini file to get SQ info
+                        psf2IniStruct = Psf2.ParseClsIniFile(iniFs);
                     }
-                }
-                else // miniPSF2
-                {                                    
-                    // unpack each lib, looking for the needed file
-                    foreach (string libPath in libPaths)
+
+                    using (FileStream fs = File.OpenRead(pPath))
                     {
-                        fileDir = Path.GetDirectoryName(libPath);
-                        fileName = Path.GetFileNameWithoutExtension(libPath);
-                        libOutputDir = Path.Combine(fileDir, fileName);
+                        Psf2 psf2File = new Psf2();
+                        psf2File.Initialize(fs, pPath);
 
-                        if (!extractedLibHash.ContainsKey(libPath))
-                        {
-                            libOutputDir = XsfUtil.UnpackPsf2(libPath, out unpkOutput, out unpkError);                                           
-                            extractedLibHash.Add(libPath, libOutputDir);
-                        }
-
-                        // look for the file in this lib
-                        sqFiles = Directory.GetFiles(libOutputDir, psf2IniStruct.SqFileName, SearchOption.AllDirectories);
+                        // check for libs
+                        libPaths = psf2File.GetLibPathArray();
+                    }
+                    if ((libPaths == null) || (libPaths.Length == 0)) // PSF2
+                    {
+                        // copy the SQ file out (should only be one)
+                        sqFiles = Directory.GetFiles(outputDir, psf2IniStruct.SqFileName, SearchOption.AllDirectories);
 
                         if (sqFiles.Length > 0)
                         {
@@ -119,18 +88,58 @@ namespace VGMToolbox.tools.xsf
                             {
                                 outputSqFileName = String.Format("{0}.SQ", outputDir);
                             }
-                            
-                            File.Copy(sqFiles[0], outputSqFileName, true);                                                                                       
-                            break;
+                            File.Copy(sqFiles[0], outputSqFileName, true);
                         }
+                    }
+                    else // miniPSF2
+                    {
+                        // unpack each lib, looking for the needed file
+                        foreach (string libPath in libPaths)
+                        {
+                            fileDir = Path.GetDirectoryName(libPath);
+                            fileName = Path.GetFileNameWithoutExtension(libPath);
+                            libOutputDir = Path.Combine(fileDir, fileName);
 
-                        // delete the unpkpsf2 output folder
-                        Directory.Delete(libOutputDir, true);
-                    
-                    } // foreach (string libPath in libPaths)
-                }
+                            if (!extractedLibHash.ContainsKey(libPath))
+                            {
+                                libOutputDir = XsfUtil.UnpackPsf2(libPath);
+                                extractedLibHash.Add(libPath, libOutputDir);
+                            }
+
+                            // look for the file in this lib
+                            sqFiles = Directory.GetFiles(libOutputDir, psf2IniStruct.SqFileName, SearchOption.AllDirectories);
+
+                            if (sqFiles.Length > 0)
+                            {
+                                if (!String.IsNullOrEmpty(psf2IniStruct.SequenceNumber))
+                                {
+                                    outputSqFileName = String.Format("{0}_n={1}.SQ", outputDir, psf2IniStruct.SequenceNumber);
+                                }
+                                else
+                                {
+                                    outputSqFileName = String.Format("{0}.SQ", outputDir);
+                                }
+
+                                File.Copy(sqFiles[0], outputSqFileName, true);
+                                break;
+                            }
+
+                            // delete the unpkpsf2 output folder
+                            if (Directory.Exists(libOutputDir))
+                            {
+                                Directory.Delete(libOutputDir, true);
+                            }
+
+                        } // foreach (string libPath in libPaths)
+                    }
+                } // if (iniFiles.Length > 0)
+
                 // delete the unpkpsf2 output folder
-                Directory.Delete(outputDir, true);
+                if (Directory.Exists(outputDir))
+                {
+                    Directory.Delete(outputDir, true);
+                }
+
             } // if (!String.IsNullOrEmpty(formatString) && (formatString.Equals(Xsf.FORMAT_NAME_PSF2)))
         }    
 
@@ -141,7 +150,10 @@ namespace VGMToolbox.tools.xsf
             // delete lib folders
             foreach (string k in extractedLibHash.Keys)
             {
-                Directory.Delete(extractedLibHash[k], true);
+                if (Directory.Exists(extractedLibHash[k]))
+                {
+                    Directory.Delete(extractedLibHash[k], true);
+                }
             }
         }        
     }
